+++ src/coq_elpi_builtins.ml
@@ -503,6 +503,27 @@ let set_accumulate_to_db, get_accumulate_to_db =
   (fun x -> f := x),
   (fun () -> !f)
 
+let argument_mode = let open Conv in let open API.AlgebraicData in declare {
+  ty = TyName "argument_mode";
+  doc = "Specify if a predicate argument is in input or output mode";
+  pp = (fun fmt _ -> Format.fprintf fmt "<todo>");
+  constructors = [
+    K("in","",N,
+        B `Input,
+        M (fun ~ok ~ko -> function `Input -> ok | _ -> ko ()));
+    K("out","",N,
+        B `Output,
+        M (fun ~ok ~ko -> function `Output -> ok | _ -> ko ()));
+  ]
+} |> CConv.(!<)
+  
+
+let set_accumulate_text_to_db, get_accumulate_text_to_db =
+  let f = ref (fun _ _ -> assert false) in
+  (fun x -> f := x),
+  (fun () -> !f)
+
+
 let class_ = let open Conv in let open API.AlgebraicData in let open Coercionops in declare {
   ty = TyName "class";
   doc = "Node of the coercion graph";
@@ -3782,6 +3803,47 @@ Supported attributes:
           f (clauses scope);
           state, (), []
      )),
+  DocAbove);
+
+  MLData argument_mode;
+
+  MLCode(Pred("coq.elpi.add-predicate",
+    In(B.string,"Db",
+    In(B.unspec B.string,"Indexing",
+    In(B.string,"PredName",
+    In(B.list (B.pair argument_mode B.string),"Spec",
+    Full(global,"Declares a new predicate PredName in the data base Db. Indexing can be left unspecified. Spec gathers a mode and a type for each argument. CAVEAT: types and indexing are strings instead of proper data types; beware parsing errors are fatal"))))),
+    (fun dbname indexing predname spec ~depth _ _ state ->
+      let dbname = Coq_elpi_utils.string_split_on_char '.' dbname in
+      let f = get_accumulate_text_to_db () in
+      let indexing =
+        match indexing with
+        | B.Given str -> ":index ("^str^") "
+        | B.Unspec -> "" in
+      let spec = spec |> List.map (fun (mode,ty) ->
+        let mode =
+          match mode with
+          | `Input -> "i:"
+          | `Output -> "o:" in
+        mode ^ "(" ^ ty ^ ")") in
+      let spec = String.concat ", " spec in
+      let text = indexing ^ "pred " ^ predname ^ " " ^ spec ^ "." in
+      f dbname text;
+      state, (), []
+      )),
+  DocAbove);
+
+  MLCode(Pred("coq.elpi.predicate",
+    In(B.string,"PredName",
+    In(list B.any,"Args",
+    Out(B.poly "prop","Pred",
+    Full(global,"Pred is the application of PredName to Args")))),
+    (fun name args _ ~depth _ _ state ->
+      let state, p = Elpi.API.Quotation.term_at ~depth state name in
+      match E.look ~depth p with
+      | Const c -> state, !: (E.mkAppL c args), []
+      | _ -> U.type_error ("predicate name expected, got " ^ name) 
+      )),
   DocAbove);
 
   LPDoc "-- Utils ------------------------------------------------------------";